+++ openid-data-classes/src/commonMain/kotlin/at/asitplus/openid/dcql/DCQLTrustedAuthorityQueryEntry.kt
@@ -1,11 +1,14 @@
 package at.asitplus.openid.dcql
 
 import at.asitplus.data.NonEmptyList
+import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
-import kotlinx.serialization.json.JsonClassDiscriminator
+import kotlinx.serialization.json.Json
+import kotlinx.serialization.json.JsonContentPolymorphicSerializer
+import kotlinx.serialization.json.JsonElement
+import kotlinx.serialization.json.JsonIgnoreUnknownKeys
 
-@Serializable
-@JsonClassDiscriminator(DCQLTrustedAuthorityQueryEntry.SerialNames.TYPE)
+@Serializable(with = DCQLTrustedAuthorityQueryEntrySerializer::class)
 sealed interface DCQLTrustedAuthorityQueryEntry {
     val type: DCQLTrustedAuthorityType
     val values: NonEmptyList<String>
@@ -15,3 +18,33 @@ sealed interface DCQLTrustedAuthorityQueryEntry {
         const val VALUES = "values"
     }
 }
+
+private class DCQLTrustedAuthorityQueryEntrySerializer :
+    JsonContentPolymorphicSerializer<DCQLTrustedAuthorityQueryEntry>(
+        DCQLTrustedAuthorityQueryEntry::class,
+    ) {
+
+    override fun selectDeserializer(
+        element: JsonElement
+    ) = selectDeserializer(
+        Json.decodeFromJsonElement(
+            DCQLTrustedAuthorityQueryEntryDeserializationDisambiguator.serializer(),
+            element
+        )
+    )
+
+    private fun selectDeserializer(
+        disambiguator: DCQLTrustedAuthorityQueryEntryDeserializationDisambiguator,
+    ) = when (disambiguator.type) {
+        DCQLTrustedAuthorityType.aki -> DCQLTrustedAuthorityQueryEntryAuthorityKeyIdentifier.serializer()
+        DCQLTrustedAuthorityType.etsi_tl -> DCQLTrustedAuthorityQueryEntryETSITrustedList.serializer()
+        DCQLTrustedAuthorityType.openid_federation -> DCQLTrustedAuthorityQueryEntryOpenIDFederation.serializer()
+    }
+}
+
+@Serializable
+@JsonIgnoreUnknownKeys
+private class DCQLTrustedAuthorityQueryEntryDeserializationDisambiguator(
+    @SerialName(DCQLTrustedAuthorityQueryEntry.SerialNames.TYPE)
+    val type: DCQLTrustedAuthorityType,
+)