+++ openid-data-classes/src/commonTest/kotlin/at/asitplus/openid/dcql/DCQLSdJwtCredentialMetadataAndValidityConstraintsTest.kt
@@ -46,7 +46,6 @@ val DCQLSdJwtCredentialMetadataAndValidityConstraintsTest by testSuite {
 
             DCQLCredentialQuery.Procedures.validateCredentialMetadataAndValidityConstraints(
                 credential = "",
-                credentialFormatIdentifier = CredentialFormatEnum.DC_SD_JWT,
                 credentialMetadataAndValidityConstraints = DCQLSdJwtCredentialMetadataAndValidityConstraints(
                     vctValues = listOf("dummy document type"),
                 ),
@@ -67,7 +66,6 @@ val DCQLSdJwtCredentialMetadataAndValidityConstraintsTest by testSuite {
         shouldThrowAny {
             DCQLCredentialQuery.Procedures.validateCredentialMetadataAndValidityConstraints(
                 credential = "",
-                credentialFormatIdentifier = CredentialFormatEnum.DC_SD_JWT,
                 credentialMetadataAndValidityConstraints = DCQLSdJwtCredentialMetadataAndValidityConstraints(
                     vctValues = listOf("dummy document type"),
                 ),