+++ openid-data-classes/src/commonTest/kotlin/at/asitplus/openid/dcql/TestCredentialQueryAdapter.kt
@@ -51,10 +51,10 @@ value class TestCredentialQueryAdapter(val dcqlQuery: DCQLQuery) {
             }
         },
         satisfiesCryptographicHolderBinding = {
-            true
+            it.satisfiesCryptographicHolderBinding
         },
         authorityKeyIdentifiers = {
-            listOf()
+            it.authorityKeyIdentifiers
         }
     )
 }